# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralix Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/licenses:
    get:
      operationId: getLicenses
      tags: [admin]
      summary: Get all licenses
      responses:
        "200":
          description: Licenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [admin]
      summary: Create license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /admin/licenses/{id}:
    delete:
      operationId: revokeLicense
      tags: [admin]
      summary: Revoke license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked

  # SUPPORT
...
components:
  schemas:
...
    License:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        plan:
          type: string
        guildId:
          type: ["string", "null"]
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, key, plan, active, createdAt]
    LicenseInput:
      type: object
      properties:
        plan:
          type: string
        guildId:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
      required: [plan]
...